+++ test/koa2/routes/index.js
@@ -1,0 +1,9 @@
+const router = require('koa-router')();
+
+router.prefix('/index');
+
+router.get('/', async (ctx, next) => {
+  ctx.body='Hi route-scanner!';
+});
+
+module.exports = router;