+++ test/koa2/routes/moduleA.js
@@ -1,0 +1,13 @@
+const router = require('koa-router')();
+
+router.prefix('/moduleA');
+
+router.get('/querySomething',async (ctx, next)=>{
+    ctx.body={
+        name:'Alex',
+        age:4,
+        gender:'male'
+    };
+});
+
+module.exports =router;