+++ docs/en/changelog.md
@@ -1,6 +1,11 @@
 # Change Log
 
 
+## v0.2.2
+Add Koa2 support.
+
+* add option `framework` to specify which framework do you chose.
+
 ## v0.2.1
 Fix the bug that registered route has double `//` as first two string.
 
@@ -11,9 +16,9 @@ changed configuration field, details below:
 * deleted options `specialMap`, `exceptMap`;
 * add options `replacePaths`: just for custom define rules to change url
   mapping.
-* add options `extraMaps`: used to add other route files path, multiple
+* add option `extraMaps`: used to add other route files path, multiple
   directories are possible;
-* add options `debug`: used to print scan logs.
+* add option `debug`: used to print scan logs.
 
 PS: after support for multi-level directories, it can be simulated as
 dynamic routing for multiple micro-services.